[pom] Update project to fork a release
will go back to original once jsoup releases, we cannot wait any longer on author to release code.  Will continue to work with the author to get items out but need this forked so we can properly fix and release formatter-maven-plugin which is broken due to jsoup and running on vulnerable code otherwise.